backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Config + validation
pydantic==2.8.2
pydantic-settings==2.3.4

# HTTP
httpx==0.27.0

# File uploads
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0

# ── AI / LLM ────────────────────────────────────────────────────────────────
# Gemini (primary — text + vision)
google-genai>=0.8.0

# ── Document extraction ──────────────────────────────────────────────────────
# OCR fallback (requires Tesseract system binary)
pytesseract==0.3.10
Pillow==10.4.0

# PDF handling
pdfminer.six==20231228
pdf2image==1.17.0          # requires poppler-utils system package
greenlet